refactor(api_wrapper): auto map mod -> module
Allows from_args to be used to compose requests containing a `module`
property. Due to `module` being a reserved word in crystal, it cannot be
used as a method argument directly. Instead of mapping each use case of
this individually this shifts this up into the macro to keep method
definitions clean.